[#11] refactor(category) - controller 전부를 index로 모음
- 카테고리 controller를 index.ts에 추가하여 controller만 import 받아도 모든 controller를 사용할 수 있도록 수정
