+++ backend/routes/category.ts
@@ -1,5 +1,5 @@
 import express from 'express'
-import controller from '../controller/category'
+import controller from '../controller'
 const router = express.Router()
 
 router.get('/', controller.getCategories)
